find and print clusters
This uses a quite dumb clustering algorithm that assumes transitively
linked files belong together. This can be quite wrong, especially
since we in practice match on only a part of the audio. For example,
with these waveforms:

 A. -----^^^^^

 B. ----------

 C. ^^^^^^^^^^

the first part of A will match B, while the second part will match C,
so B and C will be clustered together, despite having no resemblance.