print progress numbers during match
The duration of the match phase is proportional to the square of the
number of audio files, so it goes from being instantaneous with a few
files to much slower than the fingerprinting phase. When it is slow,
we want it to occasionally print a progress update, which we can
estimate with good accuracy since we know the number of comparisons.