pre-scan for short match before full search
If we don't find an OK short match, we won't find a good long match.

In the common case where the majority of files are not duplicates,
this makes the search both faster and more accurate, as the final slow
search is hardly ever run so can have a larger window.